TO-DO-LIST APPLICATION
Traditionally, we would write on a piece of paper or post it notes and act as a memory aid. As technology has evolved we have been able to create a todo lists with this website . You can use a to do list in your home and personal life, or in the workplace.